Work in progress, for now I’m working only on follow_likers.
moved min_media and max_media in set_relationship_bounds

add parameter max_relationship_ratio in set_relationship_bounds, to delimit also with a maximum ratio.

For now, this is a work in progress, so the new behavior is the default, but I’ll add a parameter to use the previous version.

Basically, first I validate the users, when i get the likers, from the likes grabbed, I add this people, in a list.
If the grabbed people is less than the desired number, pick again the likes, ad add the new ones grabbed to the list.
This process cannot be done forever, so i decided to add maximum attempts, I chose to use 3 as maximum.

Another thing, I added, a very little delay, between user validation. Now, for testing purposes is set to a very low number.
The idea, is simulate, the user checking the profile. So basically, stay a little bit on this profile, even if is not valid, and then move on to the next one

Because, we collect a large number of people to follow, I added also a delay between the follows.

For now is all, folks :)